02_s3_glue_athena/requirements.txt
pyspark
awsglue-local